+++ Mods ChoGGi/Building Placement Orientation/metadata.lua
@@ -23,6 +23,5 @@ return PlaceObj("ModDef", {
 	"lua_revision", LuaRevision or 244275,
 	"description", [[Any object you place will have the same orientation as the last placed object.
 
-Ctrl-Space to activate placement mode with the last placed object.
-Ctrl-Shift-Space to activate placement mode with the selected object (or object under mouse).]],
+Ctrl-Space to activate placement mode with the selected or last placed object (or object under mouse).]],
 })